refactor(carbon-provider): move contexts into an internal directory
Moves `TopModalContext` and `NewValidationContext` into an `__internal__` directory
within `CarbonProvider` directory.

BREAKING CHANGE: contexts have been moved to an internal directory
