+++ src/__internal__/focus-trap/focus-trap.component.tsx
@@ -17,7 +17,7 @@ import ModalContext, {
   ModalContextProps,
 } from "../../components/modal/__internal__/modal.context";
 import usePrevious from "../../hooks/__internal__/usePrevious";
-import TopModalContext from "../../components/carbon-provider/top-modal-context";
+import TopModalContext from "../../components/carbon-provider/__internal__/top-modal.context";
 
 export const TAB_GUARD_TOP = "tab-guard-top";
 export const TAB_GUARD_BOTTOM = "tab-guard-bottom";
