+++ src/__internal__/input/input-presentation.spec.tsx
@@ -11,7 +11,7 @@ import StyledInput from "./input.style";
 import sizes from "./input-sizes.style";
 import { assertStyleMatch } from "../../__spec_helper__/__internal__/test-utils";
 import { InputContext, InputGroupContext } from "../input-behaviour";
-import { NewValidationContext } from "../../components/carbon-provider/carbon-provider.component";
+import NewValidationContext from "../../components/carbon-provider/__internal__/new-validation.context";
 
 function render(props: InputPresentationProps) {
   return mount(<InputPresentation {...props} />);
