+++ src/components/carbon-provider/__internal__/top-modal-provider.component.tsx
@@ -1,24 +1,7 @@
-import React, {
-  createContext,
-  useState,
-  useEffect,
-  useRef,
-  ReactNode,
-} from "react";
+import React, { useState, useEffect, useRef, ReactNode } from "react";
+import TopModalContext from "./top-modal.context";
 
-type TopModalContextProps = {
-  topModal: HTMLElement | null;
-};
-
-const TopModalContext = createContext<TopModalContextProps>({
-  topModal: null,
-});
-
-export const TopModalContextProvider = ({
-  children,
-}: {
-  children: ReactNode;
-}) => {
+export const TopModalProvider = ({ children }: { children: ReactNode }) => {
   const [topModal, setTopModal] = useState<HTMLElement | null>(null);
 
   // can't add the setter to the global list inside useEffect because that doesn't run until
@@ -50,4 +33,4 @@ export const TopModalContextProvider = ({
   );
 };
 
-export default TopModalContext;
+export default TopModalProvider;
