+++ src/components/carbon-provider/__internal__/top-modal.context.ts
@@ -1,0 +1,9 @@
+import React from "react";
+
+type TopModalContextProps = {
+  topModal: HTMLElement | null;
+};
+
+export default React.createContext<TopModalContextProps>({
+  topModal: null,
+});
